Refactor the file uploader
