--- academy-ui/src/components/app/FileUploader.tsx
@@ -1,91 +1,0 @@
-// Replace with your actual uploadImageToS3 import
-import LoadingSpinner from '@dodao/web-core/components/core/loaders/LoadingSpinner';
-import { CreateSignedUrlInput, ImageType } from '@/graphql/generated/generated-types';
-import getBaseUrl from '@dodao/web-core/utils/api/getBaseURL';
-import { getUploadedImageUrlFromSingedUrl } from '@dodao/web-core/utils/upload/getUploadedImageUrlFromSingedUrl';
-import axios from 'axios';
-import React, { useRef, useState } from 'react';
-import styled from 'styled-components';
-
-interface Props {
-  spaceId: string;
-  objectId: string;
-  imageType: ImageType;
-  onLoading?: (loading: boolean) => void;
-  onInput?: (imageUrl: string) => void;
-  children: React.ReactNode;
-  className?: string;
-  allowedFileTypes: string[];
-}
-
-const FileSelect = styled.label`
-  input[type='file'] {
-    display: none;
-    font-weight: normal;
-  }
-  &:hover {
-    cursor: pointer;
-  }
-`;
-
-export default function FileUploader({ spaceId, objectId, imageType, onLoading, onInput, children, className, allowedFileTypes }: Props) {
-  const [loading, setLoading] = useState(false);
-  const inputRef = useRef<HTMLInputElement>(null);
-
-  async function uploadToS3AndReturnImgUrl(imageType: string, file: File, objectId: string) {
-    const input: CreateSignedUrlInput = {
-      imageType,
-      contentType: file.type,
-      objectId,
-      name: file.name.replace(' ', '_').toLowerCase(),
-    };
-
-    const response = await axios.post(`${getBaseUrl()}/api/s3-signed-urls`, { spaceId, input });
-
-    const signedUrl = response?.data?.url!;
-    await axios.put(signedUrl, file, {
-      headers: { 'Content-Type': file.type },
-    });
-
-    const imageUrl = getUploadedImageUrlFromSingedUrl(signedUrl);
-    return imageUrl;
-  }
-
-  const handleFileChange = async (e: React.ChangeEvent<HTMLInputElement>) => {
-    setLoading(true);
-    onLoading && onLoading(true);
-    const file = e.target.files![0];
-    if (!allowedFileTypes.includes(file.type)) {
-      console.log('File type not supported');
-      setLoading(false);
-      return;
-    }
-
-    try {
-      const imageUrl = await uploadToS3AndReturnImgUrl(imageType, file, objectId.replace(/[^a-z0-9]/gi, '_'));
-
-      onInput && onInput(imageUrl);
-      setLoading(false);
-      onLoading && onLoading(false);
-    } catch (error) {
-      setLoading(false);
-      onLoading && onLoading(false);
-      console.log(error);
-    }
-  };
-
-  return (
-    <div className={className}>
-      {loading ? (
-        <div className="pl-4">
-          <LoadingSpinner />
-        </div>
-      ) : (
-        <FileSelect>
-          <input type="file" ref={inputRef} onChange={handleFileChange} accept={allowedFileTypes.join(', ')} />
-          {children}
-        </FileSelect>
-      )}
-    </div>
-  );
-}
